# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio OS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface Tag {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  category?: string | null;
  createdAt: string;
}

export interface CreateTagBody {
  name: string;
  slug: string;
  /** @nullable */
  category?: string | null;
}

export interface Project {
  id: number;
  title: string;
  slug: string;
  summary: string;
  classification: string;
  status: string;
  /** @nullable */
  year?: number | null;
  isPublic: boolean;
  isFeatured: boolean;
  techStack: string[];
  tags: Tag[];
  /** @nullable */
  coverImageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectSection {
  id: number;
  projectId: number;
  type: string;
  title: string;
  content: string;
  sortOrder: number;
  createdAt: string;
}

export interface MediaAsset {
  id: number;
  projectId: number;
  type: string;
  url: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  altText?: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface ProjectDetail {
  id: number;
  title: string;
  slug: string;
  summary: string;
  classification: string;
  status: string;
  /** @nullable */
  year?: number | null;
  isPublic: boolean;
  isFeatured: boolean;
  techStack: string[];
  tags: Tag[];
  sections: ProjectSection[];
  media: MediaAsset[];
  relatedProjects: Project[];
  /** @nullable */
  coverImageUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  title: string;
  slug: string;
  summary: string;
  classification: string;
  status: string;
  /** @nullable */
  year?: number | null;
  isPublic?: boolean;
  isFeatured?: boolean;
  techStack?: string[];
  tagIds?: number[];
  /** @nullable */
  coverImageUrl?: string | null;
}

export interface UpdateProjectBody {
  title?: string;
  slug?: string;
  summary?: string;
  classification?: string;
  status?: string;
  /** @nullable */
  year?: number | null;
  isPublic?: boolean;
  isFeatured?: boolean;
  techStack?: string[];
  tagIds?: number[];
  /** @nullable */
  coverImageUrl?: string | null;
}

export interface CreateSectionBody {
  type: string;
  title: string;
  content: string;
  sortOrder: number;
}

export interface UpdateSectionBody {
  type?: string;
  title?: string;
  content?: string;
  sortOrder?: number;
}

export interface CreateMediaBody {
  type: string;
  url: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  altText?: string | null;
  sortOrder: number;
}

export interface OperatorChatBody {
  message: string;
  /** @nullable */
  conversationId?: string | null;
}

export interface ProjectCitation {
  projectId: number;
  title: string;
  slug: string;
}

export interface TourStop {
  projectId: number;
  title: string;
  slug: string;
  rationale: string;
}

export interface OperatorChatResponse {
  message: string;
  conversationId: string;
  citations: ProjectCitation[];
  tour?: TourStop[];
}

export interface OperatorConversation {
  id: number;
  conversationId: string;
  role: string;
  message: string;
  citations: ProjectCitation[];
  createdAt: string;
}

export interface VaultNote {
  id: number;
  projectId: number;
  content: string;
  createdAt: string;
  updatedAt: string;
}

export interface ExchangeMobileAuthorizationCodeBody {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface UpsertVaultNoteBody {
  content: string;
}

export interface TagCount {
  tag: Tag;
  count: number;
}

export interface ClassificationCount {
  classification: string;
  count: number;
}

export interface DashboardSummary {
  totalProjects: number;
  featuredProjects: Project[];
  recentProjects: Project[];
  tagCloud: TagCount[];
  classificationBreakdown: ClassificationCount[];
}

export type ListProjectsParams = {
  tag?: string;
  search?: string;
  classification?: string;
  isPublic?: boolean;
};