# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Portfolio OS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/media:
    get:
      operationId: listProjectMedia
      tags: [media]
      summary: List project media
      description: Get all media assets for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of media assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaAsset"
    post:
      operationId: createProjectMedia
      tags: [media]
      summary: Add media asset
      description: Add a media asset to a project (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMediaBody"
      responses:
        "201":
          description: Created media asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaAsset"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /operator/chat:
    post:
      operationId: operatorChat
      tags: [operator]
      summary: Chat with AI operator
      description: Send a message to the AI operator and receive a grounded response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OperatorChatBody"
      responses:
        "200":
          description: Operator response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperatorChatResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MediaAsset:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        type:
          type: string
        url:
          type: string
        caption:
          type: ["string", "null"]
        altText:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - type
        - url
        - sortOrder
        - createdAt
    CreateMediaBody:
      type: object
      properties:
        type:
          type: string
        url:
          type: string
        caption:
          type: ["string", "null"]
        altText:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - type
        - url
        - sortOrder
    OperatorChatBody:
      type: object
      properties:
        message:
          type: string
        conversationId:
          type: ["string", "null"]
      required:
        - message
    ProjectCitation:
      type: object
      properties:
        projectId:
          type: integer
        title:
          type: string
        slug:
          type: string
      required:
        - projectId
        - title
        - slug
    TourStop:
      type: object
      properties:
        projectId:
          type: integer
        title:
          type: string
        slug:
          type: string
        rationale:
          type: string
      required:
        - projectId
        - title
        - slug
        - rationale
    OperatorChatResponse:
      type: object
      properties:
        message:
          type: string
        conversationId:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/ProjectCitation"
        tour:
          type: array
          items:
            $ref: "#/components/schemas/TourStop"
      required:
        - message
        - conversationId
        - citations
...