# lib/api-zod/src/generated/types/mediaAsset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Portfolio OS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MediaAsset {
  id: number;
  projectId: number;
  type: string;
  url: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  altText?: string | null;
  sortOrder: number;
  createdAt: Date;
}